Fix type mismatch: brandColor hex literals need .toInt() for Int type

+++ healthconnect_to_gcal/hc_export_app/app/src/main/kotlin/com/hcexport/SourceBrands.kt
@@ -12,17 +12,17 @@ object SourceBrands {
     )
 
     private val knownBrands = listOf(
-        BrandSpec("samsung", "Samsung Health", R.drawable.ic_source_samsung_health, 0xFF0D99FF),
-        BrandSpec("fitbit",  "Fitbit",          R.drawable.ic_source_fitbit,         0xFF00B0B9),
-        BrandSpec("garmin",  "Garmin Connect",  R.drawable.ic_source_garmin,         0xFF0082C3),
-        BrandSpec("polar",   "Polar Flow",      R.drawable.ic_source_polar,          0xFFE60000),
-        BrandSpec("strava",  "Strava",          R.drawable.ic_source_strava,         0xFFFC4C02),
-        BrandSpec("whoop",   "WHOOP",           R.drawable.ic_source_whoop,          0xFF00ADEF),
-        BrandSpec("circular","Circular",        R.drawable.ic_source_circular,       0xFF1A1A1A),
-        BrandSpec("lyfta",   "Lyfta",           R.drawable.ic_source_lyfta,          0xFF8B5CF6),
-        BrandSpec("withings","Withings",        R.drawable.ic_source_withings,       0xFF4A90D9),
-        BrandSpec("huawei",  "Huawei Health",   R.drawable.ic_source_huawei,         0xFFCF0A2C),
-        BrandSpec("fitness", "Google Fit",      R.drawable.ic_source_google_fit,     0xFF4285F4),
+        BrandSpec("samsung", "Samsung Health", R.drawable.ic_source_samsung_health, 0xFF0D99FF.toInt()),
+        BrandSpec("fitbit",  "Fitbit",          R.drawable.ic_source_fitbit,         0xFF00B0B9.toInt()),
+        BrandSpec("garmin",  "Garmin Connect",  R.drawable.ic_source_garmin,         0xFF0082C3.toInt()),
+        BrandSpec("polar",   "Polar Flow",      R.drawable.ic_source_polar,          0xFFE60000.toInt()),
+        BrandSpec("strava",  "Strava",          R.drawable.ic_source_strava,         0xFFFC4C02.toInt()),
+        BrandSpec("whoop",   "WHOOP",           R.drawable.ic_source_whoop,          0xFF00ADEF.toInt()),
+        BrandSpec("circular","Circular",        R.drawable.ic_source_circular,       0xFF1A1A1A.toInt()),
+        BrandSpec("lyfta",   "Lyfta",           R.drawable.ic_source_lyfta,          0xFF8B5CF6.toInt()),
+        BrandSpec("withings","Withings",        R.drawable.ic_source_withings,       0xFF4A90D9.toInt()),
+        BrandSpec("huawei",  "Huawei Health",   R.drawable.ic_source_huawei,         0xFFCF0A2C.toInt()),
+        BrandSpec("fitness", "Google Fit",      R.drawable.ic_source_google_fit,     0xFF4285F4.toInt()),
     )
 
     fun forPackage(pkg: String): BrandInfo {
@@ -31,7 +31,7 @@ object SourceBrands {
             BrandInfo(spec.displayName, spec.iconResId, spec.brandColor)
         } else {
             val fallback = pkg.substringAfterLast('.').replaceFirstChar { it.uppercase() }
-            BrandInfo(fallback, R.drawable.ic_source_generic, 0xFF64748B)
+            BrandInfo(fallback, R.drawable.ic_source_generic, 0xFF64748B.toInt())
         }
     }
 
@@ -43,6 +43,6 @@ object SourceBrands {
         val keyword: String,
         val displayName: String,
         val iconResId: Int,
-        val brandColor: Long,
+        val brandColor: Int,
     )
 }
